backend/requirements.txt
fastapi[all]>=0.115.0
uvicorn[standard]>=0.25.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
alembic>=1.13.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8
pydantic>=2.5.0
pydantic-settings>=2.1.0
celery[redis]>=5.3.4
redis>=5.0.1
httpx>=0.26.0

# WebSocket support
websockets>=12.0

# AI/ML
openai>=1.3.0
google-generativeai>=0.3.0

# PDF generation
reportlab>=4.0.0
Pillow>=10.1.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0

# Security
cryptography>=41.0.0

# Email
aiosmtplib>=1.3.0